feat(buck): extend the command links with ability to set the output format
Add the flag --format. The flag --format accepts a string to define the output format.
Add support for the format 'json'. If the format is set to json the link struct is printed as parsable JSON object.
Add the flag --json. The flag --json sets the flag --format to json

Signed-off-by: Merzough Münker <mmuenker@digitaix.com>