feat(buck): extend the command links with ability to set the output format (#465)
* feat(buck): extend the command links with ability to set the output format

Add the flag --format. The flag --format accepts a string to define the output format.
Add support for the format 'json'. If the format is set to json the link struct is printed as parsable JSON object.
Add the flag --json. The flag --json sets the flag --format to json

Signed-off-by: Merzough Münker <mmuenker@digitaix.com>

* fix(buck): use proper flag declaration function

Signed-off-by: Merzough Münker <mmuenker@digitaix.com>

* fix(buck): use a enum instead of a generic strings to represent the select format type

Signed-off-by: Merzough Münker <mmuenker@digitaix.com>

* fix(buck): remove --json flag from the links command

Signed-off-by: Merzough Münker <mmuenker@digitaix.com>

Co-authored-by: Sander Pick <sanderpick@gmail.com>